refactor(#31): useUserStatus(useLoginStatus) 타입 세부 지정
