+++ src/components/organisms/PublishSideBar/index.tsx
@@ -2,7 +2,7 @@ import { toast } from 'react-toastify';
 import { ButtonVariant, ColorVariant } from '@/constants';
 import { ModalKey } from '@/constants/keys';
 import useOpenClose from '@/hooks/useOpenClose';
-import useUserStatus from '@/hooks/useLoginStatus';
+import useUserStatus from '@/hooks/useUserStatus';
 import { Book } from '@/types/features/book';
 import { BookReviewId } from '@/types/features/bookReview';
 import bookReviewStore from '@/stores/bookReviewStore';
